Add redirection to login page
