+++ lib/withAuth.js
@@ -1,18 +1,6 @@
-import Router from 'next/router';
 import React from 'react';
 import nookies from 'nookies';
-
-const redirect = (ctx = {}, target) => {
-  if (ctx.res) {
-    // server side redirection
-    // More info: https://bit.ly/2TcSlQ9
-    ctx.res.writeHead(303, { Location: target });
-    ctx.res.end();
-  } else {
-    // client side redirection
-    Router.push(target);
-  }
-};
+import redirect from './redirect';
 
 // HOC
 export default (ComposedComponent) => class WithAuth extends React.Component {
@@ -23,7 +11,7 @@ export default (ComposedComponent) => class WithAuth extends React.Component {
       composedInitialProps = await ComposedComponent.getInitialProps(ctx);
     }
 
-    // Need to check if the token is still valid, API
+    // Need to check if the token is still valid (hasn't expired), with an API endpoint?
     const cookies = nookies.get(ctx);
 
     if (!cookies.token) {
